feat: per-language voice phrases, AX terminal focus, polish + refactor
Voice phrasing
- Per-event phrase pools (TEMPLATES_RESPONSE / TEMPLATES_NOTIFICATION) in
  phrases/{en,fr,hi,it,pt}.sh, copied from stackone-say-hooks. notify.sh
  picks one at random per nudge and formats it with the project name
  (basename of $PWD with hyphens/underscores split, common acronyms
  expanded for clarity)
- voice_to_lang / voice_to_kokoro_lang map the configured Kokoro voice
  prefix to a phrase-file language and a stackvox --lang code, so a
  French voice (ff_*) gets French phrasing
- Voice "replaces" sound: when STACKNUDGE_VOICE=true, the chime is
  suppressed across all paths (banner, banner-off, frontmost-suppression
  fallback) so we don't double-cue. Notifier treats an empty --sound as
  "no sound"

VS Code / Cursor terminal-pane focus
- Before sending the approval keystroke, AppActivator walks the editor's
  AX tree and focuses a terminal pane labelled with the agent's binary
  name (claude/gemini/codex). With shell integration on (default) the
  tab title reflects the running command so the match is reliable for
  one-CC-per-window. Falls back gracefully if no match
- Fixed regression where canvas drilling was matching the editor input
  area instead of xterm. Reverted to AX-press of the matched tab; route
  Enter via `tell process "Code"` so the keystroke can't accidentally
  land on a momentarily-frontmost other app

Hotkey UX
- Default flipped from cmd+shift+n (collides with "New Incognito Window"
  in browsers, "New Window" in many editors, "New Folder" in Finder) to
  cmd+opt+n
- Settings has a recordable Hotkey row at the top: ⏎ enters record mode,
  next combo with at least one modifier commits and re-registers the
  global hotkey live. ↑↓/Tab during recording cancel and navigate so
  users who entered by mistake aren't trapped. If the OS rejects the
  combo (already bound), an inline red caption surfaces it
- Hotkey.encode reverses parse so NSEvent modifiers + keyCode round-trip
  back into the spec format

Config file watcher
- DispatchSourceFileSystemObject on ~/.stack-nudge/config; edits via
  "Open config file…" or any external editor flow back into the running
  panel without a restart. Re-arms on .rename/.delete with a 100ms delay
  so atomic-save editors don't orphan the watcher

Permissions window
- Fixed crash on open: collectionBehavior had .canJoinAllSpaces and
  .moveToActiveSpace simultaneously (mutually exclusive — AppKit threw
  NSInternalInconsistencyException). Sync construction was also masking
  the throw, so opening from the panel handler appeared as a hang. Now
  async-dispatched + .moveToActiveSpace only
- Window level bumped to .modalPanel so it floats above the panel itself
  and System Settings, letting users grant both permissions in one pass

Tabs + UI polish
- Header tab strip with bell icon + Events / Sessions / Settings pills
  and active-count badges; one combined ⌘1-3 keycap on the right edge
  instead of per-tab keycaps
- ScrollViews use ThinScrollers (NSViewRepresentable) to shrink the
  scroller to .mini — about half the default width
- Voice toggle preview speaks "Voice notifications enabled"; voice cycle
  in Settings previews the new voice with a random conversational phrase

Sessions
- Inline rename: select a row, press n, edit in the row's TextField,
  ⏎ commits / esc cancels. Footer hints flip to match recording state
- Throbber + "Checking for sessions…" empty state until first scan
  completes
- Agent detection broadened to node/deno/bun-hosted binaries (gemini-cli
  installed via npm shows comm=node; we match args)

Refactor
- panel/PanelNav.swift owns navigation state + settings model (PanelMode,
  SettingsActions, PanelNav). Out of Settings.swift since all pages read
  it
- panel/Components.swift hosts shared SwiftUI helpers (FooterHint,
  KeyCapView, FooterDivider, ThinScrollers, PageFooter)
- panel/Speaker.swift extracts the stackvox-say invocation so menu-bar
  toggles and voice cycle previews share one path
- ~150 lines lighter across the page files; each page now declares a
  focused PageFooter { … } instead of duplicating styling

Misc
- make reload also copies notify.sh + phrases/ into ~/.stack-nudge so
  hook-side edits propagate the same cycle as the daemon binary
- make dev watches notify.sh and phrases/ in addition to Swift sources
- README updated to cover tabs, sessions, settings, hotkey recorder,
  voice phrasing, and the new dev workflow
