+++ panel/PanelNav.swift
@@ -1,0 +1,220 @@
+import AppKit
+import SwiftUI
+
+enum PanelMode {
+    case events
+    case sessions
+    case settings
+}
+
+// Action callbacks the controller wires into nav so settings rows like
+// "Check permissions" / "Open config" / "Quit" can fire effects without the
+// SwiftUI view needing to know about windows or app-level state.
+struct SettingsActions {
+    let checkPermissions: () -> Void
+    let openConfig:       () -> Void
+    let quit:             () -> Void
+}
+
+// Owns the panel's navigation state plus the live values the Settings page
+// renders and mutates. Lives outside Settings.swift because the navigation
+// model (mode, selected row, hotkey state) is read across all pages — only
+// the settings field set is settings-page-specific.
+final class PanelNav: ObservableObject {
+
+    @Published var mode: PanelMode = .events
+    @Published var selectedSettingIndex: Int = 0
+
+    @Published var hotkeyDisplay:   String = "cmd+opt+n"
+    @Published var recordingHotkey: Bool = false
+    @Published var hotkeyError:     String?
+    @Published var bannerEnabled:   Bool = true
+    @Published var voiceEnabled:    Bool = false
+    @Published var soundStop:       String = "Glass"
+    @Published var soundPermission: String = "Ping"
+    @Published var voice:           String = "af_heart"
+    @Published var voiceSpeed:      Double = 1.1
+    @Published var voicesAvailable: [String] = []
+    @Published var voicesLoading:   Bool = true
+
+    var actions: SettingsActions?
+    // Wired by PanelController so nav can re-register the global hotkey
+    // without owning the Hotkey instance directly. Returns true if the
+    // new spec registered successfully.
+    var setHotkey: ((String) -> Bool)?
+
+    static let macSounds = [
+        "Basso", "Blow", "Bottle", "Frog", "Funk", "Glass", "Hero",
+        "Morse", "Ping", "Pop", "Purr", "Sosumi", "Submarine", "Tink",
+    ]
+
+    static let speedMin = 0.60
+    static let speedMax = 1.60
+    static let speedStep = 0.05
+
+    // Conversational, agent-flavoured preview phrases. Picked at random each
+    // time the user cycles voices so they hear varied cadence and intonation
+    // rather than the same word over and over.
+    static let voicePreviewPhrases = [
+        "Oh, I have a question.",
+        "Could you take a look at this?",
+        "I need your approval to continue.",
+        "Just finished that task — want to review?",
+        "Hey, I've got something to show you.",
+        "Quick question before I move on.",
+        "Hmm, this one's tricky. Mind helping?",
+        "All done. Ready when you are.",
+        "Should I proceed with this change?",
+        "I'd love your input on this.",
+    ]
+
+    var rowCount: Int { 10 }
+
+    // Row layout (kept in one place so the controller, view, and indexing
+    // logic all agree on what each row index means):
+    //   0  Hotkey                hotkey-record
+    //   1  Banner notifications  toggle
+    //   2  Voice notifications   toggle
+    //   3  Agent done sound      cycle
+    //   4  Permission sound      cycle
+    //   5  Voice                 cycle
+    //   6  Speed                 cycle
+    //   7  Check permissions…    action
+    //   8  Open config file…     action
+    //   9  Quit panel            action
+
+    // MARK: - Disk I/O
+
+    func loadFromConfig() {
+        let config = ConfigFile.read()
+        hotkeyDisplay   = config["STACKNUDGE_PANEL_HOTKEY"]    ?? "cmd+opt+n"
+        bannerEnabled   = ConfigFile.bool(config, "STACKNUDGE_BANNER", default: true)
+        voiceEnabled    = ConfigFile.bool(config, "STACKNUDGE_VOICE",  default: false)
+        soundStop       = config["STACKNUDGE_SOUND_STOP"]       ?? "Glass"
+        soundPermission = config["STACKNUDGE_SOUND_PERMISSION"] ?? "Ping"
+        voice           = config["STACKNUDGE_VOICE_NAME"]       ?? "af_heart"
+        voiceSpeed      = Double(config["STACKNUDGE_VOICE_SPEED"] ?? "") ?? 1.1
+    }
+
+    func loadVoices() {
+        Task.detached(priority: .userInitiated) {
+            let names = Self.runStackvoxVoices()
+            await MainActor.run { [weak self] in
+                self?.voicesAvailable = names
+                self?.voicesLoading = false
+            }
+        }
+    }
+
+    private nonisolated static func runStackvoxVoices() -> [String] {
+        let stackvox = "\(NSHomeDirectory())/.stack-nudge/venv/bin/stackvox"
+        guard FileManager.default.isExecutableFile(atPath: stackvox) else { return [] }
+        let task = Process()
+        task.executableURL = URL(fileURLWithPath: stackvox)
+        task.arguments = ["voices"]
+        let pipe = Pipe()
+        task.standardOutput = pipe
+        task.standardError = Pipe()
+        do { try task.run() } catch { return [] }
+        let data = pipe.fileHandleForReading.readDataToEndOfFile()
+        task.waitUntilExit()
+        return (String(data: data, encoding: .utf8) ?? "")
+            .split(separator: "\n")
+            .map { $0.trimmingCharacters(in: .whitespaces) }
+            .filter { !$0.isEmpty && !$0.contains(" ") }
+    }
+
+    // MARK: - Row movement
+
+    func selectNextRow() {
+        guard rowCount > 0 else { return }
+        selectedSettingIndex = (selectedSettingIndex + 1) % rowCount
+    }
+
+    func selectPrevRow() {
+        guard rowCount > 0 else { return }
+        selectedSettingIndex = (selectedSettingIndex - 1 + rowCount) % rowCount
+    }
+
+    // MARK: - Cycle / activate
+
+    // Enter: toggles flip, cycle rows step forward, actions fire, hotkey
+    // row enters record mode.
+    func activate() {
+        switch selectedSettingIndex {
+        case 0: startRecordingHotkey()
+        case 7: actions?.checkPermissions()
+        case 8: actions?.openConfig()
+        case 9: actions?.quit()
+        default: applyCycle(forward: true)
+        }
+    }
+
+    func cycleForward()  { applyCycle(forward: true) }
+    func cycleBackward() { applyCycle(forward: false) }
+
+    private func applyCycle(forward: Bool) {
+        switch selectedSettingIndex {
+        case 0:
+            // Cycle on the hotkey row also enters record mode.
+            startRecordingHotkey()
+        case 1:
+            bannerEnabled.toggle()
+            ConfigFile.write(key: "STACKNUDGE_BANNER", value: bannerEnabled ? "true" : "false")
+        case 2:
+            voiceEnabled.toggle()
+            ConfigFile.write(key: "STACKNUDGE_VOICE", value: voiceEnabled ? "true" : "false")
+        case 3:
+            soundStop = step(soundStop, in: Self.macSounds, forward: forward, key: "STACKNUDGE_SOUND_STOP", preview: true)
+        case 4:
+            soundPermission = step(soundPermission, in: Self.macSounds, forward: forward, key: "STACKNUDGE_SOUND_PERMISSION", preview: true)
+        case 5:
+            guard !voicesLoading, !voicesAvailable.isEmpty else { return }
+            voice = step(voice, in: voicesAvailable, forward: forward, key: "STACKNUDGE_VOICE_NAME", preview: false)
+            let phrase = Self.voicePreviewPhrases.randomElement() ?? "Hello."
+            Speaker.speak(phrase, voice: voice, speed: String(format: "%.2f", voiceSpeed))
+        case 6:
+            let next = forward ? voiceSpeed + Self.speedStep : voiceSpeed - Self.speedStep
+            voiceSpeed = max(Self.speedMin, min(Self.speedMax, (next * 100).rounded() / 100))
+            ConfigFile.write(key: "STACKNUDGE_VOICE_SPEED", value: String(format: "%.2f", voiceSpeed))
+        default:
+            break
+        }
+    }
+
+    private func step(_ current: String, in list: [String], forward: Bool, key: String, preview: Bool) -> String {
+        guard !list.isEmpty else { return current }
+        let idx = list.firstIndex(of: current) ?? 0
+        let next = forward ? (idx + 1) % list.count : (idx - 1 + list.count) % list.count
+        let value = list[next]
+        ConfigFile.write(key: key, value: value)
+        if preview { NSSound(named: NSSound.Name(value))?.play() }
+        return value
+    }
+
+    // MARK: - Hotkey recording
+
+    func startRecordingHotkey() {
+        recordingHotkey = true
+        hotkeyError = nil
+    }
+
+    func cancelRecordingHotkey() {
+        recordingHotkey = false
+    }
+
+    func commitHotkey(_ spec: String) {
+        // Try to register first; only persist if it took. Failing means the
+        // combo was rejected by the OS (already-registered global, malformed,
+        // etc.) — keep the previous one and surface a message to the user.
+        guard let setHotkey, setHotkey(spec) else {
+            hotkeyError = "‘\(spec)’ is unavailable — already bound by another app"
+            recordingHotkey = false
+            return
+        }
+        hotkeyError = nil
+        hotkeyDisplay = spec
+        ConfigFile.write(key: "STACKNUDGE_PANEL_HOTKEY", value: spec)
+        recordingHotkey = false
+    }
+}
